Refactor cli/index.js
