+++ packages/cli/index.js
@@ -1,27 +1,13 @@
 #!/usr/bin/env node
 
 // Entry file for MarkBind project
-const chokidar = require('chokidar');
-const fs = require('fs-extra');
-const path = require('path');
 const program = require('commander');
-const Promise = require('bluebird');
 
-const _ = {};
-_.isBoolean = require('lodash/isBoolean');
-
-const { Site } = require('@markbind/core');
-const { pageVueServerRenderer } = require('@markbind/core/src/Page/PageVueServerRenderer');
-
-const fsUtil = require('@markbind/core/src/utils/fsUtil');
-const {
-  INDEX_MARKDOWN_FILE,
-  LAZY_LOADING_SITE_FILE_NAME,
-} = require('@markbind/core/src/Site/constants');
-
-const liveServer = require('./src/lib/live-server');
-const cliUtil = require('./src/util/cliUtil');
 const logger = require('./src/util/logger');
+const { build } = require('./src/cmd/build');
+const { deploy } = require('./src/cmd/deploy');
+const { init } = require('./src/cmd/init');
+const { serve } = require('./src/cmd/serve');
 
 const CLI_VERSION = require('./package.json').version;
 
@@ -33,11 +19,6 @@ process.stdout.write(
 function printHeader() {
   logger.logo();
   logger.log(` v${CLI_VERSION}`);
-}
-
-function handleError(error) {
-  logger.error(error.message);
-  process.exitCode = 1;
 }
 
 program
@@ -59,29 +40,7 @@ program
   .alias('i')
   .description('init a markbind website project')
   .action((root, options) => {
-    const rootFolder = path.resolve(root || process.cwd());
-    if (options.convert) {
-      if (fs.existsSync(path.resolve(rootFolder, 'site.json'))) {
-        logger.error('Cannot convert an existing MarkBind website!');
-        return;
-      }
-    }
-    Site.initSite(rootFolder, options.template)
-      .then(() => {
-        logger.info('Initialization success.');
-      })
-      .then(() => {
-        if (options.convert) {
-          logger.info('Converting to MarkBind website.');
-          const outputRoot = path.join(rootFolder, '_site');
-          new Site(rootFolder, outputRoot).convert()
-            .then(() => {
-              logger.info('Conversion success.');
-            })
-            .catch(handleError);
-        }
-      })
-      .catch(handleError);
+    init(root, options);
   });
 
 program
@@ -99,197 +58,7 @@ program
   .option('-d, --dev', 'development mode, enabling live & hot reload for frontend source files.')
   .description('build then serve a website from a directory')
   .action((userSpecifiedRoot, options) => {
-    if (options.dev) {
-      logger.useDebugConsole();
-    }
-
-    let rootFolder;
-    try {
-      rootFolder = cliUtil.findRootFolder(userSpecifiedRoot, options.siteConfig);
-
-      if (options.forceReload && options.onePage) {
-        handleError(new Error('Oops! You shouldn\'t need to use the --force-reload option with --one-page.'));
-        process.exit();
-      }
-    } catch (err) {
-      handleError(err);
-    }
-    const logsFolder = path.join(rootFolder, '_markbind/logs');
-    const outputFolder = path.join(rootFolder, '_site');
-
-    const presentDefaultFile = fsUtil.fileExists(INDEX_MARKDOWN_FILE) ? INDEX_MARKDOWN_FILE : false;
-    if (options.onePage === true && !presentDefaultFile) {
-      handleError(new Error('Oops! It seems that you didn\'t have the default file index.md.'));
-      process.exit();
-    }
-    let onePagePath = options.onePage === true ? presentDefaultFile : options.onePage;
-    onePagePath = onePagePath ? fsUtil.ensurePosix(onePagePath) : onePagePath;
-
-    const reloadAfterBackgroundBuild = () => {
-      logger.info('All opened pages will be reloaded.');
-      liveServer.reloadActiveTabs();
-    };
-
-    const site = new Site(rootFolder, outputFolder, onePagePath,
-                          options.forceReload, options.siteConfig, options.dev,
-                          options.backgroundBuild, reloadAfterBackgroundBuild);
-
-    const syncOpenedPages = () => {
-      logger.info('Synchronizing opened pages list before reload');
-      const normalizedActiveUrls = liveServer.getActiveUrls().map((url) => {
-        const completeUrl = path.extname(url) === '' ? path.join(url, 'index') : url;
-        return fsUtil.removeExtension(completeUrl);
-      });
-      site.changeCurrentOpenedPages(normalizedActiveUrls);
-    };
-
-    const addHandler = (filePath) => {
-      logger.info(`[${new Date().toLocaleTimeString()}] Reload for file add: ${filePath}`);
-      if (onePagePath) {
-        syncOpenedPages();
-      }
-      Promise.resolve('').then(async () => {
-        if (site.isFilepathAPage(filePath) || site.isDependencyOfPage(filePath)) {
-          return site.rebuildSourceFiles(filePath);
-        }
-        return site.buildAsset(filePath);
-      }).catch((err) => {
-        logger.error(err.message);
-      });
-    };
-
-    const changeHandler = (filePath) => {
-      logger.info(`[${new Date().toLocaleTimeString()}] Reload for file change: ${filePath}`);
-      if (onePagePath) {
-        syncOpenedPages();
-      }
-      Promise.resolve('').then(async () => {
-        if (path.basename(filePath) === path.basename(site.siteConfigPath)) {
-          return site.reloadSiteConfig();
-        }
-        if (site.isDependencyOfPage(filePath)) {
-          return site.rebuildAffectedSourceFiles(filePath);
-        }
-        return site.buildAsset(filePath);
-      }).catch((err) => {
-        logger.error(err.message);
-      });
-    };
-
-    const removeHandler = (filePath) => {
-      logger.info(`[${new Date().toLocaleTimeString()}] Reload for file deletion: ${filePath}`);
-      if (onePagePath) {
-        syncOpenedPages();
-      }
-      Promise.resolve('').then(async () => {
-        if (site.isFilepathAPage(filePath) || site.isDependencyOfPage(filePath)) {
-          return site.rebuildSourceFiles(filePath);
-        }
-        return site.removeAsset(filePath);
-      }).catch((err) => {
-        logger.error(err.message);
-      });
-    };
-
-    // server config
-    const serverConfig = {
-      open: options.open,
-      logLevel: 0,
-      root: outputFolder,
-      port: options.port || 8080,
-      middleware: [],
-      mount: [],
-    };
-
-    site
-      .readSiteConfig()
-      .then(async (config) => {
-        serverConfig.mount.push([config.baseUrl || '/', outputFolder]);
-
-        if (options.dev) {
-          // eslint-disable-next-line global-require
-          const webpackDevConfig = require('@markbind/core-web/webpack.dev');
-
-          await webpackDevConfig.serverEntry(pageVueServerRenderer.updateMarkBindVueBundle, rootFolder);
-
-          const getMiddlewares = webpackDevConfig.clientEntry;
-          getMiddlewares(`${config.baseUrl}/markbind`)
-            .forEach(middleware => serverConfig.middleware.push(middleware));
-        }
-
-        if (onePagePath) {
-          const lazyReloadMiddleware = function (req, res, next) {
-            const urlExtension = path.posix.extname(req.url);
-
-            const hasEndingSlash = req.url.endsWith('/');
-            const hasNoExtension = urlExtension === '';
-            const isHtmlFileRequest = urlExtension === '.html' || hasEndingSlash || hasNoExtension;
-
-            if (!isHtmlFileRequest || req.url.endsWith('._include_.html')) {
-              next();
-              return;
-            }
-
-            if (hasNoExtension && !hasEndingSlash) {
-              // Urls of type 'host/userGuide' - check if 'userGuide' is a raw file or does not exist
-              const diskFilePath = path.resolve(rootFolder, req.url);
-              if (!fs.existsSync(diskFilePath) || !(fs.statSync(diskFilePath).isDirectory())) {
-                // Request for a raw file
-                next();
-                return;
-              }
-            }
-
-            const urlWithoutBaseUrl = req.url.replace(config.baseUrl, '');
-            // Map 'hostname/userGuide/' and 'hostname/userGuide' to hostname/userGuide/index.
-            const urlWithIndex = (hasNoExtension || hasEndingSlash)
-              ? path.posix.join(urlWithoutBaseUrl, 'index')
-              : urlWithoutBaseUrl;
-            const urlWithoutExtension = fsUtil.removeExtension(urlWithIndex);
-
-            const didInitiateRebuild = site.changeCurrentPage(urlWithoutExtension);
-            if (didInitiateRebuild) {
-              req.url = fsUtil.ensurePosix(path.join(config.baseUrl || '/', LAZY_LOADING_SITE_FILE_NAME));
-            }
-            next();
-          };
-
-          const onePageHtmlUrl = `${config.baseUrl}/${onePagePath.replace(/\.md$/, '.html')}`;
-          serverConfig.open = serverConfig.open && onePageHtmlUrl;
-
-          serverConfig.middleware.push(lazyReloadMiddleware);
-        } else {
-          serverConfig.open = serverConfig.open && `${config.baseUrl}/`;
-        }
-
-        return site.generate();
-      })
-      .then(() => {
-        const watcher = chokidar.watch(rootFolder, {
-          ignored: [
-            logsFolder,
-            outputFolder,
-            /(^|[/\\])\../,
-            x => x.endsWith('___jb_tmp___'), x => x.endsWith('___jb_old___'), // IDE temp files
-          ],
-          ignoreInitial: true,
-        });
-        watcher
-          .on('add', addHandler)
-          .on('change', changeHandler)
-          .on('unlink', removeHandler);
-      })
-      .then(() => {
-        const server = liveServer.start(serverConfig);
-        server.addListener('listening', () => {
-          const address = server.address();
-          const serveHost = address.address === '0.0.0.0' ? '127.0.0.1' : address.address;
-          const serveURL = `http://${serveHost}:${address.port}`;
-          logger.info(`Serving "${outputFolder}" at ${serveURL}`);
-          logger.info('Press CTRL+C to stop ...');
-        });
-      })
-      .catch(handleError);
+    serve(userSpecifiedRoot, options);
   });
 
 program
@@ -300,22 +69,7 @@ program
   .option('-s, --site-config <file>', 'specify the site config file (default: site.json)')
   .description('build a website')
   .action((userSpecifiedRoot, output, options) => {
-    // if --baseUrl contains no arguments (options.baseUrl === true) then set baseUrl to empty string
-    const baseUrl = _.isBoolean(options.baseUrl) ? '' : options.baseUrl;
-    let rootFolder;
-    try {
-      rootFolder = cliUtil.findRootFolder(userSpecifiedRoot, options.siteConfig);
-    } catch (err) {
-      handleError(err);
-    }
-    const defaultOutputRoot = path.join(rootFolder, '_site');
-    const outputFolder = output ? path.resolve(process.cwd(), output) : defaultOutputRoot;
-    new Site(rootFolder, outputFolder, undefined, undefined, options.siteConfig)
-      .generate(baseUrl)
-      .then(() => {
-        logger.info('Build success!');
-      })
-      .catch(handleError);
+    build(userSpecifiedRoot, output, options);
   });
 
 program
@@ -326,36 +80,7 @@ program
   .option('-s, --site-config <file>', 'specify the site config file (default: site.json)')
   .description('deploy the latest build of the site to the repo\'s Github pages')
   .action((userSpecifiedRoot, options) => {
-    let rootFolder;
-    try {
-      rootFolder = cliUtil.findRootFolder(userSpecifiedRoot, options.siteConfig);
-    } catch (err) {
-      handleError(err);
-    }
-    const outputFolder = path.join(rootFolder, '_site');
-
-    // Choose to build or not build depending on --no-build flag
-    // We cannot chain generate and deploy while calling generate conditionally, so we split with if-else
-    const site = new Site(rootFolder, outputFolder, undefined, undefined, options.siteConfig);
-    if (options.build) {
-      site.generate()
-        .then(() => {
-          logger.info('Build success!');
-        })
-        .then(() => {
-          site.deploy(options.ci)
-            .then(depUrl => (depUrl !== null ? logger.info(
-              `The website has been deployed at: ${depUrl}`)
-              : logger.info('Deployed!')));
-        })
-        .catch(handleError);
-    } else {
-      site.deploy(options.ci)
-        .then(depUrl => (depUrl !== null ? logger.info(
-          `The website has been deployed at: ${depUrl}`)
-          : logger.info('Deployed!')))
-        .catch(handleError);
-    }
+    deploy(userSpecifiedRoot, options);
   });
 
 program.parse(process.argv);
